Rewrite data storage handling
Instead of repository, cached/backup storage we now have a backend which
in turn has very simple functions for loading and saving (when
applicable) data from and to backends such as cache, backup (file
system) and http.

And instead of us telling unleash to "use cache if stale", "don't
ask server for data" we instead implement a LoaderStrategy where we get
passed the backend container. It is now up to us to decide if we want to
fetch from server, use cache, use filesystem, handle locking updates
from server and so on...